pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "design-review-env"
version = "2.0.0"
description = "AI-Driven Engineering Design Review — Multi-domain agentic RL environment"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "BSD-3-Clause"}
authors = [
    {name = "jetty"},
]
keywords = ["openenv", "reinforcement-learning", "engineering", "design-review", "cad", "ai-agent"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: BSD License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "openenv-core>=0.1.0",
    "pydantic>=2.0",
    "numpy>=1.24.0",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-asyncio",
]

[project.urls]
Homepage = "https://github.com/jetty/design-review-env"
Repository = "https://github.com/jetty/design-review-env"
Issues = "https://github.com/jetty/design-review-env/issues"

[tool.hatch.build.targets.wheel]
packages = ["design_review_env"]
